Fixed 9776 - PhoneTextBox - Text color after disabling/re-enabling control
The problem here is more serious than the issue title lets on. This wasn't just a case of the text foreground color being off; the problem here is that the common state that should've been called "Normal" was actually called "Enabled" in the PhoneTextBox's default template. As a result, the "Normal" state can never be entered since it doesn't exist, and disabling the control causes it to permanently get placed into the disabled visual state, in which a TextBox with white text exists in the foreground.

The fix for this is simple enough: I just renamed "Enabled" to "Normal". While I was at it, I also added this state to the list of visual states in the TemplateVisualState attribute collection and removed the setting of the visibility of the HintBorder from the Normal state - this happens automatically just when the animations associated with the Disabled state stop when we return to the Normal state, so there's no reason to make it explicit.

Work Items: 9776