Fixes for LoopingSelector, including:
7855: LoopingSelector SelectedItem Issue
http://silverlight.codeplex.com/workitem/7855

10458: LoopingSelector performance improvement
http://silverlight.codeplex.com/workitem/10458

LoopingSelector realizes items when it’s initialized (before it’s expanded). The fix is to do so only when IsExpanded is true. This improves initialization performance.

7946: Primitives.LoopingSelector implementation of DataSource property
http://silverlight.codeplex.com/workitem/7946

The XAML parser bypasses the getter/setter of vanilla CLR properties. Thus, it’s not a good place to run event subscription.

9477: LoopingSelector ItemTemplate property updating
http://silverlight.codeplex.com/workitem/9477

When the ItemTemplate changes, it didn’t have any effect. So I added an UpdateItemTemplates method.

9508: LoopingSelector missing items
http://silverlight.codeplex.com/workitem/9508

This is the same issue as the recycling problem for bug 7855.

8706: WP7: LoopingSelector Brake and AnimatePanel functions
http://silverlight.codeplex.com/workitem/8706

This is a bit of theoretical bug as I wasn’t able to repro it. I made sure though that the factor is always positive.

Work Items: 7855,10458,7946,9477,9508,8706