Fixed 8396 - TransitionService - Immediately pressing back after a navigation results in a crash.
Let’s say you have page1 and page2 with some navigation transitions on them. If you navigate from page1 to page2 and then have page2 immediately call NavigationService.GoBack(), the app will crash.
The issue is that the transition animation for page1 -> page2 is still running when GoBack is called and  page1 is still a children of a content presenter (TransitionFrame has 2 content presenters). And since the visual tree can’t have the same element in two places, you can’t set the content of TransitionFrame to page1 yet.
The fix is to make sure that before starting a new navigation (when OnNavigating is called), that the running transition, if any, completes and that the transition states are reset. That’s exactly what EnsureLastTransitionIsComplete does.

Work Item: 8396